safari-600.7-branch/LayoutTests/fast/dom/Geolocation/reentrant-success-expected.txt

Tests that reentrant calls to Geolocation methods from the success callback are OK.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy
PASS successfullyParsed is true

TEST COMPLETE
